+++ Braintree/src/main/java/com/braintreepayments/api/BraintreeFragment.java
@@ -610,6 +610,15 @@ public class BraintreeFragment extends BrowserSwitchFragment {
         return Collections.unmodifiableList(mCachedPaymentMethodNonces);
     }
 
+    /**
+     * Check if this Braintree fragment is still active.
+     *
+     * @return {@code true} if still active and process can proceed, {@code false} otherwise.
+     */
+    public boolean isActive() {
+        return isAdded();
+    }
+
     public void sendAnalyticsEvent(final String eventFragment) {
         final AnalyticsEvent request = new AnalyticsEvent(mContext, getSessionId(), mIntegrationType, eventFragment);
         waitForConfiguration(new ConfigurationListener() {